+++ python/nano/src/bigdl/nano/deps/openvino/pytorch/pytorch_openvino_model.py
@@ -16,7 +16,7 @@
 import os
 from ..core.openvino_model import OpenVINOModel
 from ...inference.pytorch.model import AcceleratedLightningModule
-from .utility import export
+from .pytorch_openvino_utils import export
 import torch
